requirements.txt
fastapi
python-dotenv
uvicorn
websockets
gunicorn
aiohttp
openai